+++ src/scicloj/clay/v2/page.clj
@@ -67,7 +67,21 @@
                         "https://code.jquery.com/ui/1.13.1/jquery-ui.min.js"]}}
    :md-default {:js {:from-local-copy
                      ["https://code.jquery.com/jquery-3.6.0.min.js"
-                      "https://code.jquery.com/ui/1.13.1/jquery-ui.min.js"]}}})
+                      "https://code.jquery.com/ui/1.13.1/jquery-ui.min.js"]}}
+   :htmlwidgets-ggplotly {:js {:from-local-copy-of-repo
+                               [{:gh-repo "scicloj/ggplotly-deps"
+                                 :relative-path ""
+                                 :paths ["lib/htmlwidgets-1.6.2/htmlwidgets.js"
+                                         "lib/plotly-binding-4.10.4.9000/plotly.js"
+                                         "lib/typedarray-0.1/typedarray.min.js"
+                                         "lib/jquery-3.5.1/jquery.min.js"
+                                         "lib/crosstalk-1.2.1/js/crosstalk.min.js"
+                                         "lib/plotly-main-2.11.1/plotly-latest.min.js"]}]}
+                          :css {:from-local-copy-of-repo
+                                [{:gh-repo "scicloj/ggplotly-deps"
+                                  :relative-path ""
+                                  :paths ["lib/crosstalk-1.2.1/css/crosstalk.min.css"
+                                          "lib/plotly-htmlwidgets-css-2.11.1/plotly-htmlwidgets.css"]}]}}})
 
 (def include
   {:js hiccup.page/include-js