+++ src/scicloj/clay/v2/prepare.clj
@@ -444,3 +444,7 @@
 (add-preparer-from-value-fn!
  :kind/observable
  #'item/observable)
+
+(add-preparer-from-value-fn!
+ :kind/htmlwidgets-ggplotly
+ #'item/ggplotly)