+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file. This change
 - page generation - minor refactoring
 - updated Kindly version: `"4-alpha18"`
 - support for JS & CSS deps from github repos
+- `kind/htmlwidgets-ggplotly` support - WIP
 
 ## [2-alpha84] - 2024-03-16
 - bugfix in `kind/table` preparation: dissocing irrelevant details correctly